Moved EbolaRiskMap to 'components' directory
