+++ src/components/EbolaRiskMap/__tests__/EbolaRiskMap.test.js
